+++ app/src/main/java/woowacourse/movie/view/reservation/result/ReservationCompleteActivity.kt
@@ -4,18 +4,14 @@ import android.content.Context
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
-import android.widget.TextView
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import androidx.databinding.BindingAdapter
 import woowacourse.movie.data.mapper.toDomain
 import woowacourse.movie.databinding.ActivityReservationCompleteBinding
 import woowacourse.movie.view.dialog.DialogFactory
 import woowacourse.movie.view.reservation.TicketUi
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
 
 class ReservationCompleteActivity : AppCompatActivity(), ReservationCompleteContract.View {
     private val presenter by lazy {
@@ -81,14 +77,4 @@ class ReservationCompleteActivity : AppCompatActivity(), ReservationCompleteCont
             Intent(context, ReservationCompleteActivity::class.java)
                 .putExtra(KEY_TICKET, ticketUi)
     }
-}
-
-@BindingAdapter("dateTime")
-fun setDateTime(
-    view: TextView,
-    date: LocalDateTime,
-) {
-    val formatter = DateTimeFormatter.ofPattern("yyyy.M.d. HH:mm")
-    val dateTimeFormat = date.format(formatter)
-    view.text = dateTimeFormat.toString()
 }
